backend/parsing-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
alembic==1.12.1
redis==5.0.1
celery==5.3.6
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2 